Refactor auth, authz, and the is_admin migration for exadmin
