+++ lib/pairmotron/plug/require_authentication.ex
@@ -1,16 +1,4 @@
-defimpl ExAdmin.Authentication, for: Plug.Conn do
-  alias Pairmotron.Router.Helpers
-  alias Pairmotron.Plug.Authenticate, as: Auth
-
-  def use_authentication?(_), do: true
-  def current_user(conn), do: Auth.current_user(conn)
-  def current_user_name(conn), do: Auth.current_user(conn).name
-  def session_path(conn, :destroy), do: Helpers.session_path(conn, :delete)
-  def session_path(conn, action), do: Helpers.session_path(conn, action)
-end
-
-
-defmodule Pairmotron.Plug.Authenticate do
+defmodule Pairmotron.Plug.RequireAuthentication do
   alias Pairmotron.Router.Helpers, as: Routes
   import Plug.Conn
 
@@ -43,5 +31,4 @@ defmodule Pairmotron.Plug.Authenticate do
     |> halt
   end
 
- def current_user(conn), do: conn.assigns[:current_user] || Guardian.Plug.current_resource(conn)
 end
