+++ web/admin/authenticate.ex
@@ -1,0 +1,10 @@
+defimpl ExAdmin.Authentication, for: Plug.Conn do
+  alias Pairmotron.Router.Helpers
+  alias Pairmotron.Authentication, as: Auth
+
+  def use_authentication?(_), do: true
+  def current_user(conn), do: Auth.current_user(conn)
+  def current_user_name(conn), do: Auth.current_user(conn).name
+  def session_path(conn, :destroy), do: Helpers.session_path(conn, :delete)
+  def session_path(conn, action), do: Helpers.session_path(conn, action)
+end
